Implement unicode-dependent parts of the spec
This adds a script to download and compress unicode data for blocks and
categories. Packed data is stored as json and integrated into the built
module by rollup.

All features should now be implemented, so this also removes the check
whether a given pattern is supported, and the possible null return value
associated with unsupported patterns.